requirements.txt
pandas
numpy
matplotlib
scikit-learn
scipy
streamlit
pillow